+++ cyber_stad_tool_v1/view/option.py
@@ -1,0 +1,7 @@
+from abc import ABC
+
+class Option(ABC):
+    def reconnaissance_option(self):
+        print("This is reconnaisance option");
+    def exploit_option(self):
+        print("This is exploit option")
